Re-organising modules and creating terratorch/models/backbones/utils.py
