+++ terratorch/models/backbones/__init__.py
@@ -6,4 +6,4 @@ import terratorch.models.backbones.prithvi_vit
 import terratorch.models.backbones.scalemae
 
 from terratorch.models.backbones.unet import UNet
-
+import terratorch.models.backbones.utils
