+++ terratorch/models/backbones/unet.py
@@ -7,350 +7,12 @@ from torch.nn import Upsample as Upsample
 import torch.nn as nn
 import torch.utils.checkpoint as cp
 
-from terratorch.models.decoders.utils import ConvModule
-
 from torch.nn.modules.batchnorm import _BatchNorm
 from torch.nn.modules.instancenorm import _InstanceNorm
 SyncBatchNorm_ = nn.SyncBatchNorm
 
-
-def build_upsample_layer(cfg: Dict, **kwargs) -> nn.Module:
-    """Build upsample layer.
-
-    Args:
-        cfg (dict): The upsample layer config, which should contain:
-
-            - type (str): Layer type.
-            - scale_factor (int): Upsample ratio, which is not applicable to
-              deconv.
-            - layer args: Args needed to instantiate a upsample layer.
-        args (argument list): Arguments passed to the ``__init__``
-            method of the corresponding conv layer.
-        kwargs (keyword arguments): Keyword arguments passed to the
-            ``__init__`` method of the corresponding conv layer.
-
-    Returns:
-        nn.Module: Created upsample layer.
-    """
-    upsample = Upsample
-
-    upsample_layer = upsample(**cfg)
-    conv_layer = ConvModule(
-        **kwargs,
-        kernel_size=3,
-        stride=1,
-        padding=1,)
-
-    layer = nn.Sequential(upsample_layer, conv_layer)
-
-    return layer
-
-class UpConvBlock(nn.Module):
-    """Upsample convolution block in decoder for UNet.
-
-    This upsample convolution block consists of one upsample module
-    followed by one convolution block. The upsample module expands the
-    high-level low-resolution feature map and the convolution block fuses
-    the upsampled high-level low-resolution feature map and the low-level
-    high-resolution feature map from encoder.
-
-    Args:
-        conv_block (nn.Sequential): Sequential of convolutional layers.
-        in_channels (int): Number of input channels of the high-level
-        skip_channels (int): Number of input channels of the low-level
-        high-resolution feature map from encoder.
-        out_channels (int): Number of output channels.
-        num_convs (int): Number of convolutional layers in the conv_block.
-            Default: 2.
-        stride (int): Stride of convolutional layer in conv_block. Default: 1.
-        dilation (int): Dilation rate of convolutional layer in conv_block.
-            Default: 1.
-        with_cp (bool): Use checkpoint or not. Using checkpoint will save some
-            memory while slowing down the training speed. Default: False.
-        conv_cfg (dict | None): Config dict for convolution layer.
-            Default: None.
-        norm_cfg (dict | None): Config dict for normalization layer.
-            Default: dict(type='BN').
-        act_cfg (dict | None): Config dict for activation layer in ConvModule.
-            Default: dict(type='ReLU').
-        upsample_cfg (dict): The upsample config of the upsample module in
-            decoder. Default: dict(type='InterpConv'). If the size of
-            high-level feature map is the same as that of skip feature map
-            (low-level feature map from encoder), it does not need upsample the
-            high-level feature map and the upsample_cfg is None.
-        dcn (bool): Use deformable convolution in convolutional layer or not.
-            Default: None.
-        plugins (dict): plugins for convolutional layers. Default: None.
-    """
-
-    def __init__(self,
-                 conv_block,
-                 in_channels,
-                 skip_channels,
-                 out_channels,
-                 num_convs=2,
-                 stride=1,
-                 dilation=1,
-                 with_cp=False,
-                 conv_cfg=None,
-                 norm_cfg=dict(type='BN'),
-                 act_cfg=dict(type='ReLU'),
-                 upsample_cfg=None,
-                 dcn=None,
-                 plugins=None):
-        super(UpConvBlock, self).__init__()
-        assert dcn is None, 'Not implemented yet.'
-        assert plugins is None, 'Not implemented yet.'
-
-        self.conv_block = conv_block(
-            in_channels=2 * skip_channels,
-            out_channels=out_channels,
-            num_convs=num_convs,
-            stride=stride,
-            dilation=dilation,
-            with_cp=with_cp,
-            conv_cfg=conv_cfg,
-            norm_cfg=norm_cfg,
-            act_cfg=act_cfg,
-            dcn=None,
-            plugins=None)
-        # TODO Upsampling as alternative for transposed convolution
-
-        if upsample_cfg is not None:
-            self.upsample = build_upsample_layer(
-                cfg=upsample_cfg, in_channels=in_channels, out_channels=skip_channels) 
-        else:
-            self.upsample = ConvModule(
-                in_channels,
-                skip_channels,
-                **self._default_conv_config,)
-                #TODO add more options for configuring these layers
-                #conv_cfg=conv_cfg,
-                #norm_cfg=norm_cfg,
-                #act_cfg=act_cfg)
-
-    @property
-    def _default_conv_config(self):
-
-        return {"kernel_size": 4,
-                "stride": 1,
-                "padding": 1,
-                "transpose": True,
-                "scale_factor": 2}
-
-    def forward(self, skip, x):
-        """Forward function."""
-        x = self.upsample(x)
-        out = torch.cat([skip, x], dim=1)
-        out = self.conv_block(out)
-        return out
-    
-class BasicConvBlock(nn.Module):
-    """Basic convolutional block for UNet.
-
-    This module consists of several plain convolutional layers.
-
-    Args:
-        in_channels (int): Number of input channels.
-        out_channels (int): Number of output channels.
-        num_convs (int): Number of convolutional layers. Default: 2.
-        stride (int): Whether use stride convolution to downsample
-            the input feature map. If stride=2, it only uses stride convolution
-            in the first convolutional layer to downsample the input feature
-            map. Options are 1 or 2. Default: 1.
-        dilation (int): Whether use dilated convolution to expand the
-            receptive field. Set dilation rate of each convolutional layer and
-            the dilation rate of the first convolutional layer is always 1.
-            Default: 1.
-        with_cp (bool): Use checkpoint or not. Using checkpoint will save some
-            memory while slowing down the training speed. Default: False.
-        conv_cfg (dict | None): Config dict for convolution layer.
-            Default: None.
-        norm_cfg (dict | None): Config dict for normalization layer.
-            Default: dict(type='BN').
-        act_cfg (dict | None): Config dict for activation layer in ConvModule.
-            Default: dict(type='ReLU').
-        dcn (bool): Use deformable convolution in convolutional layer or not.
-            Default: None.
-        plugins (dict): plugins for convolutional layers. Default: None.
-    """
-
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 num_convs=2,
-                 stride=1,
-                 dilation=1,
-                 with_cp=False,
-                 conv_cfg=None,
-                 norm_cfg=dict(type='BN'),
-                 act_cfg=dict(type='ReLU'),
-                 dcn=None,
-                 plugins=None):
-        super(BasicConvBlock, self).__init__()
-        assert dcn is None, 'Not implemented yet.'
-        assert plugins is None, 'Not implemented yet.'
-
-        self.with_cp = with_cp
-        convs = []
-        for i in range(num_convs):
-            convs.append(
-                ConvModule(
-                    in_channels=in_channels if i == 0 else out_channels,
-                    out_channels=out_channels,
-                    kernel_size=3,
-                    stride=stride if i == 0 else 1,
-                    dilation=1 if i == 0 else dilation,
-                    padding=1 if i == 0 else dilation,))
-                    # TODO add more options for configuring these layers
-                    #conv_cfg=conv_cfg,
-                    #norm_cfg=norm_cfg,
-                    #act_cfg=act_cfg))
-
-        self.convs = nn.Sequential(*convs)
-
-    def forward(self, x):
-        """Forward function."""
-
-        if self.with_cp and x.requires_grad:
-            out = cp.checkpoint(self.convs, x)
-        else:
-            out = self.convs(x)
-        return out
-
-
-class DeconvModule(nn.Module):
-    """Deconvolution upsample module in decoder for UNet (2X upsample).
-
-    This module uses deconvolution to upsample feature map in the decoder
-    of UNet.
-
-    Args:
-        in_channels (int): Number of input channels.
-        out_channels (int): Number of output channels.
-        with_cp (bool): Use checkpoint or not. Using checkpoint will save some
-            memory while slowing down the training speed. Default: False.
-        norm_cfg (dict | None): Config dict for normalization layer.
-            Default: dict(type='BN').
-        act_cfg (dict | None): Config dict for activation layer in ConvModule.
-            Default: dict(type='ReLU').
-        kernel_size (int): Kernel size of the convolutional layer. Default: 4.
-    """
-
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 with_cp=False,
-                 norm_cfg=dict(type='BN'),
-                 act_cfg=dict(type='ReLU'),
-                 *,
-                 kernel_size=4,
-                 scale_factor=2):
-        super(DeconvModule, self).__init__()
-
-        assert (kernel_size - scale_factor >= 0) and\
-               (kernel_size - scale_factor) % 2 == 0,\
-               f'kernel_size should be greater than or equal to scale_factor '\
-               f'and (kernel_size - scale_factor) should be even numbers, '\
-               f'while the kernel size is {kernel_size} and scale_factor is '\
-               f'{scale_factor}.'
-
-        stride = scale_factor
-        padding = (kernel_size - scale_factor) // 2
-        self.with_cp = with_cp
-        deconv = nn.ConvTranspose2d(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding)
-
-        norm_name, norm = build_norm_layer(norm_cfg, out_channels)
-        activate = build_activation_layer(act_cfg)
-        self.deconv_upsamping = nn.Sequential(deconv, norm, activate)
-
-    def forward(self, x):
-        """Forward function."""
-
-        if self.with_cp and x.requires_grad:
-            out = cp.checkpoint(self.deconv_upsamping, x)
-        else:
-            out = self.deconv_upsamping(x)
-        return out
-
-
-class InterpConv(nn.Module):
-    """Interpolation upsample module in decoder for UNet.
-
-    This module uses interpolation to upsample feature map in the decoder
-    of UNet. It consists of one interpolation upsample layer and one
-    convolutional layer. It can be one interpolation upsample layer followed
-    by one convolutional layer (conv_first=False) or one convolutional layer
-    followed by one interpolation upsample layer (conv_first=True).
-
-    Args:
-        in_channels (int): Number of input channels.
-        out_channels (int): Number of output channels.
-        with_cp (bool): Use checkpoint or not. Using checkpoint will save some
-            memory while slowing down the training speed. Default: False.
-        norm_cfg (dict | None): Config dict for normalization layer.
-            Default: dict(type='BN').
-        act_cfg (dict | None): Config dict for activation layer in ConvModule.
-            Default: dict(type='ReLU').
-        conv_cfg (dict | None): Config dict for convolution layer.
-            Default: None.
-        conv_first (bool): Whether convolutional layer or interpolation
-            upsample layer first. Default: False. It means interpolation
-            upsample layer followed by one convolutional layer.
-        kernel_size (int): Kernel size of the convolutional layer. Default: 1.
-        stride (int): Stride of the convolutional layer. Default: 1.
-        padding (int): Padding of the convolutional layer. Default: 1.
-        upsample_cfg (dict): Interpolation config of the upsample layer.
-            Default: dict(
-                scale_factor=2, mode='bilinear', align_corners=False).
-    """
-
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 with_cp=False,
-                 norm_cfg=dict(type='BN'),
-                 act_cfg=dict(type='ReLU'),
-                 *,
-                 conv_cfg=None,
-                 conv_first=False,
-                 kernel_size=1,
-                 stride=1,
-                 padding=0,
-                 upsample_cfg=dict(
-                     scale_factor=2, mode='bilinear', align_corners=False)):
-        super(InterpConv, self).__init__()
-
-        self.with_cp = with_cp
-        conv = ConvModule(
-            in_channels,
-            out_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            conv_cfg=conv_cfg,
-            norm_cfg=norm_cfg,
-            act_cfg=act_cfg)
-        upsample = Upsample(**upsample_cfg)
-        if conv_first:
-            self.interp_upsample = nn.Sequential(conv, upsample)
-        else:
-            self.interp_upsample = nn.Sequential(upsample, conv)
-
-    def forward(self, x):
-        """Forward function."""
-
-        if self.with_cp and x.requires_grad:
-            out = cp.checkpoint(self.interp_upsample, x)
-        else:
-            out = self.interp_upsample(x)
-        return out
-
+from terratorch.models.backbones.utils import UpConvBlock, BasicConvBlock
+from terratorch.models.decoders.utils import ConvModule
 
 class UNet(nn.Module):
     """UNet backbone.
